% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creepyalien-package.R
\docType{package}
\name{creepyalien-package}
\alias{creepyalien-package}
\title{creepyalien: Puzzle Game for the R Console}
\description{
Puzzle game that can be played in the R console. Help the alien to find the ship.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rolkra/creepyalien}
}

}
\author{
\strong{Maintainer}: Roland Krasser \email{roland.krasser@gmail.com}

}
\keyword{internal}
