% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creepyalien.R
\name{creepyalien}
\alias{creepyalien}
\title{Play creepyalien in the R-console}
\usage{
creepyalien(animate = TRUE, sound = TRUE)
}
\arguments{
\item{animate}{Animated CLI graphics? (TRUE|FALSE)}

\item{sound}{Play sound? (TRUE|FALSE)}
}
\value{
Nothing
}
\description{
Play creepyalien in the R-console
}
\examples{
## Start game (in interactive R sessions)
if (interactive())  {
   creepyalien()
}
}
